+++ Tests/TuistGeneratorTests/Generator/Mocks/MockInfoPlistContentProvider.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+@testable import TuistGenerator
+
+final class MockInfoPlistContentProvider: InfoPlistContentProviding {
+    var contentArgs: [(target: Target, extendedWith: [String: InfoPlist.Value])] = []
+    var contentStub: [String: Any]?
+
+    func content(target: Target, extendedWith: [String: InfoPlist.Value]) -> [String: Any]? {
+        contentArgs.append((target: target, extendedWith: extendedWith))
+        return contentStub ?? [:]
+    }
+}