+++ CHANGELOG.md
@@ -4,7 +4,12 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 
 ## Next
 
+### Added
+
+- New InfoPlist type, `.extendingDefault([:])` https://github.com/tuist/tuist/pull/448 by @pepibumur
+
 ### Fixed
+
 - Transitively link static dependency's dynamic dependencies correctly https://github.com/tuist/tuist/pull/484 by @adamkhazi
 - Prevent embedding static frameworks https://github.com/tuist/tuist/pull/490 by @kwridan